Test y modularización graficas y funciones datos close #48
