+++ .travis.yml
@@ -4,4 +4,4 @@ python:
 install:
   - pip3 install -r requirements.txt
 script:
-  - pytest ./test/
+  - pytest ./test/ ./test_integration/
