+++ covid_reports.py
@@ -1,9 +1,7 @@
 import logging
 import os
 import sys
-import pandas as pd
-from datetime import date, timedelta, datetime
-import matplotlib.pyplot as plt
+from graphics import *
 from telegram import InlineKeyboardButton, InlineKeyboardMarkup, ReplyKeyboardMarkup, KeyboardButton
 from telegram.ext import Updater, CommandHandler, ConversationHandler, MessageHandler, Filters, CallbackQueryHandler
 
@@ -46,28 +44,7 @@ INFO_ESPANA_100_DEATH_MEDIA, INFO_ESPANA_ALL = range(130)
 mode = os.getenv("MODE")
 TOKEN = os.getenv("TOKEN")
 
-# Database used in proyect
-url_casos = 'https://raw.githubusercontent.com/datadista/datasets/master/COVID%2019/ccaa_covid19_datos_isciii_nueva_serie.csv'
-util_cols = ['fecha', 'ccaa', 'num_casos']
-df_ccaa_casos = pd.read_csv(url_casos, sep=',', usecols=util_cols)
-
-url_muertes = 'https://raw.githubusercontent.com/datadista/datasets/master/COVID%2019/ccaa_covid19_fallecidos_por_fecha_defuncion_nueva_serie_original.csv'
-df_ccaa_muertes = pd.read_csv(url_muertes, sep=',')
-
-url_hospital = 'https://raw.githubusercontent.com/datadista/datasets/master/COVID%2019/ccaa_ingresos_camas_convencionales_uci.csv'
-util_cols = ['Fecha', 'CCAA', 'Total Pacientes COVID ingresados', '% Camas Ocupadas COVID',
-             'Total pacientes COVID en UCI', '% Camas Ocupadas UCI COVID', 'Ingresos COVID últimas 24 h',
-             'Altas COVID últimas 24 h']
-df_ccaa_hospital = pd.read_csv(url_hospital, sep=',', usecols=util_cols)
-
-url_edad = 'https://raw.githubusercontent.com/datadista/datasets/master/COVID%2019/nacional_covid19_rango_edad.csv'
-util_cols = ['fecha', 'rango_edad', 'sexo', 'casos_confirmados', 'fallecidos']
-df_edad = pd.read_csv(url_edad, sep=',', usecols=util_cols)
-
-df_ccaa_habitantes = pd.read_csv('./data/habitantes_ccaa.csv', sep=',')
-
-
-global current_state, conv_handler
+global current_state #, conv_handler
 
 if mode == "dev":
     def run(updater):
@@ -1984,25 +1961,8 @@ def usuario_pulsa_boton_anterior(update, context):
         parse_mode='HTML'
     )
 
-
-def normalize(s):
-    replacements = (
-        ("á", "a"),
-        ("é", "e"),
-        ("í", "i"),
-        ("ó", "o"),
-        ("ú", "u"),
-        (" ", ""),
-        (".", ""),
-        ("ñ", "n"),
-    )
-    for a, b in replacements:
-        s = s.replace(a, b).replace(a.upper(), b.upper())
-    return s
-
-
 def main():
-    global conv_handler
+    # global conv_handler
 
     logger.info("Starting bot")
     updater = Updater(TOKEN)
@@ -4224,688 +4184,6 @@ def main():
 
     run(updater)
 
-
-########################  FUNCIONALIDAD BBDD  ########################
-def format_date(fecha):
-    return datetime.strptime(fecha, "%Y-%m-%d").strftime("%d-%m-%Y")
-
-
-def casos_acumulados(current_autonomy):
-    return str(df_ccaa_casos.groupby(['ccaa'])['num_casos'].sum()[current_autonomy])
-
-
-def incremento_ultimo_dia(current_autonomy):
-    df_loc = df_ccaa_casos.loc[(df_ccaa_casos['fecha'] == fecha_actualizacion(current_autonomy)) &
-                               (df_ccaa_casos['ccaa'] == current_autonomy)]
-    return str(df_loc['num_casos'].values[0])
-
-
-def media_casos_semana(current_autonomy):
-    ultima_fecha = fecha_actualizacion(current_autonomy)
-    fecha = datetime.strptime(ultima_fecha, "%Y-%m-%d")
-
-    total = 0
-    for i in range(7):
-        fecha2 = fecha - timedelta(days=i)
-        fecha_semana_antes = fecha2.strftime("%Y-%m-%d")
-
-        df_loc = df_ccaa_casos.loc[(df_ccaa_casos['fecha'] == fecha_semana_antes) &
-                                   (df_ccaa_casos['ccaa'] == current_autonomy)]
-        total += int(df_loc['num_casos'].values[0])
-
-    return str(round(total/7, 1))
-
-
-def fecha_actualizacion(current_autonomy):
-    actual_day = date.today()
-    dias_antes = 0
-    day_before = timedelta(days=dias_antes)
-
-    # La primera fecha de la que se tiene datos
-    while df_ccaa_casos.loc[df_ccaa_casos['fecha'] == str(actual_day-day_before)].empty:
-        dias_antes += 1
-        day_before = timedelta(days=dias_antes)
-
-    # Los ultimos datos de la ccaa
-    df_loc = df_ccaa_casos.loc[(df_ccaa_casos['fecha'] == str(actual_day-day_before)) &
-                               (df_ccaa_casos['ccaa'] == current_autonomy)]
-
-    while df_loc['num_casos'].values[0] == 0:
-        dias_antes += 1
-        day_before = timedelta(days=dias_antes)
-        df_loc = df_ccaa_casos.loc[(df_ccaa_casos['fecha'] == str(actual_day-day_before)) &
-                                   (df_ccaa_casos['ccaa'] == current_autonomy)]
-
-    return str(actual_day-day_before)
-
-
-def muertes_totales(current_autonomy):
-    return str(df_ccaa_muertes.groupby(['CCAA'])['Fallecidos'].sum()[current_autonomy])
-
-
-def muertes_ultimo_dia(current_autonomy):
-    df_loc = df_ccaa_muertes.loc[(df_ccaa_muertes['Fecha'] == fecha_actualizacion_muertes()) &
-                                 (df_ccaa_muertes['CCAA'] == current_autonomy)]
-    return str(df_loc['Fallecidos'].values[0])
-
-
-def media_muertes_semana(current_autonomy):
-    ultima_fecha = fecha_actualizacion_muertes()
-    fecha = datetime.strptime(ultima_fecha, "%Y-%m-%d")
-
-    total = 0
-    for i in range(7):
-        fecha2 = fecha - timedelta(days=i)
-        fecha_semana_antes = fecha2.strftime("%Y-%m-%d")
-
-        df_loc = df_ccaa_muertes.loc[(df_ccaa_muertes['Fecha'] == fecha_semana_antes) &
-                                     (df_ccaa_muertes['CCAA'] == current_autonomy)]
-        total += int(df_loc['Fallecidos'].values[0])
-
-    return str(round(total/7, 1))
-
-
-def tasa_letalidad(current_autonomy):
-    return str(round(int(muertes_totales(current_autonomy))*100/int(casos_acumulados(current_autonomy)), 2))
-
-
-def fecha_actualizacion_muertes():
-    actual_day = date.today()
-    dias_antes = 0
-    day_before = timedelta(days=dias_antes)
-
-    # La primera fecha de la que se tiene datos
-    while df_ccaa_muertes.loc[df_ccaa_muertes['Fecha'] == str(actual_day - day_before)].empty:
-        dias_antes += 1
-        day_before = timedelta(days=dias_antes)
-
-    return str(actual_day-day_before)
-
-
-def pacientes_ingresados(current_autonomy):
-    df_loc = df_ccaa_hospital.loc[(df_ccaa_hospital['Fecha'] == fecha_actualizacion_hospital(current_autonomy)) &
-                                  (df_ccaa_hospital['CCAA'] == current_autonomy)]
-    return str(int(df_loc['Total Pacientes COVID ingresados']))
-
-
-def porcentaje_camas_ocupadas(current_autonomy):
-    df_loc = df_ccaa_hospital.loc[(df_ccaa_hospital['Fecha'] == fecha_actualizacion_hospital(current_autonomy)) &
-                                  (df_ccaa_hospital['CCAA'] == current_autonomy)]
-    return str(df_loc['% Camas Ocupadas COVID'].values[0])
-
-
-def pacientes_ingresados_uci(current_autonomy):
-    df_loc = df_ccaa_hospital.loc[(df_ccaa_hospital['Fecha'] == fecha_actualizacion_hospital(current_autonomy)) &
-                                  (df_ccaa_hospital['CCAA'] == current_autonomy)]
-    return str(int(df_loc['Total pacientes COVID en UCI']))
-
-
-def porcentaje_camas_uci_ocupadas(current_autonomy):
-    df_loc = df_ccaa_hospital.loc[(df_ccaa_hospital['Fecha'] == fecha_actualizacion_hospital(current_autonomy)) &
-                                  (df_ccaa_hospital['CCAA'] == current_autonomy)]
-    return str(df_loc['% Camas Ocupadas UCI COVID'].values[0])
-
-
-def ingresados_ultimo_dia(current_autonomy):
-    df_loc = df_ccaa_hospital.loc[(df_ccaa_hospital['Fecha'] == fecha_actualizacion_hospital(current_autonomy)) &
-                                  (df_ccaa_hospital['CCAA'] == current_autonomy)]
-    return str(int(df_loc['Ingresos COVID últimas 24 h']))
-
-
-def altas_ultimo_dia(current_autonomy):
-    df_loc = df_ccaa_hospital.loc[(df_ccaa_hospital['Fecha'] == fecha_actualizacion_hospital(current_autonomy)) &
-                                  (df_ccaa_hospital['CCAA'] == current_autonomy)]
-    return str(int(df_loc['Altas COVID últimas 24 h']))
-
-
-def fecha_actualizacion_hospital(current_autonomy):
-    actual_day = date.today()
-    dias_antes = 0
-    day_before = timedelta(days=dias_antes)
-
-    # La primera fecha de la que se tiene datos
-    while df_ccaa_hospital.loc[df_ccaa_hospital['Fecha'] == str(actual_day-day_before)].empty:
-        dias_antes += 1
-        day_before = timedelta(days=dias_antes)
-
-    # Los ultimos datos de la ccaa
-    df_loc = df_ccaa_hospital.loc[(df_ccaa_hospital['Fecha'] == str(actual_day-day_before)) &
-                                  (df_ccaa_hospital['CCAA'] == current_autonomy)]
-
-    while df_loc['Total Pacientes COVID ingresados'].values[0] == 0:
-        dias_antes += 1
-        day_before = timedelta(days=dias_antes)
-        df_loc = df_ccaa_hospital.loc[(df_ccaa_hospital['Fecha'] == str(actual_day-day_before)) &
-                                      (df_ccaa_hospital['CCAA'] == current_autonomy)]
-
-    return str(actual_day-day_before)
-
-
-def casos_acumulados_espana():
-    return str(df_ccaa_casos['num_casos'].sum())
-
-
-def incremento_ultimo_dia_espana():
-    df_loc = df_ccaa_casos.loc[df_ccaa_casos['fecha'] == fecha_actualizacion_espana()]
-    return str(df_loc['num_casos'].sum())
-
-
-def media_casos_semana_espana():
-    ultima_fecha = fecha_actualizacion_espana()
-    fecha = datetime.strptime(ultima_fecha, "%Y-%m-%d")
-
-    total = 0
-    for i in range(7):
-        fecha2 = fecha - timedelta(days=i)
-        fecha_semana_antes = fecha2.strftime("%Y-%m-%d")
-
-        df_loc = df_ccaa_casos.loc[df_ccaa_casos['fecha'] == fecha_semana_antes]
-        total += int(df_loc['num_casos'].sum())
-
-    return str(round(total / 7, 1))
-
-
-def fecha_actualizacion_espana():
-    actual_day = date.today()
-    dias_antes = 0
-    day_before = timedelta(days=dias_antes)
-
-    # La primera fecha de la que se tiene datos
-    while df_ccaa_casos.loc[df_ccaa_casos['fecha'] == str(actual_day - day_before)].empty:
-        dias_antes += 1
-        day_before = timedelta(days=dias_antes)
-
-    return str(actual_day-day_before)
-
-
-def casos_edad(edad):
-    df_loc = df_edad.loc[(df_edad['fecha'] == fecha_actualizacion_edad()) & (df_edad['sexo'] == 'ambos') & (df_edad['rango_edad'] == edad)]
-    return str(df_loc['casos_confirmados'].values[0])
-
-
-def muertos_edad(edad):
-    df_loc = df_edad.loc[(df_edad['fecha'] == fecha_actualizacion_edad()) & (df_edad['sexo'] == 'ambos') & (df_edad['rango_edad'] == edad)]
-    return str(df_loc['fallecidos'].values[0])
-
-
-def tasa_edad(edad):
-    return str(round(int(muertos_edad(edad))*100/int(casos_edad(edad)), 2))
-
-
-def total_casos_edad():
-    df_loc = df_edad.loc[(df_edad['fecha'] == fecha_actualizacion_edad()) & (df_edad['sexo'] == 'ambos')]
-    return str(df_loc['casos_confirmados'].sum())
-
-
-def total_muertos_edad():
-    df_loc = df_edad.loc[(df_edad['fecha'] == fecha_actualizacion_edad()) & (df_edad['sexo'] == 'ambos')]
-    return str(df_loc['fallecidos'].sum())
-
-
-def total_tasa_edad():
-    return str(round(int(total_muertos_edad())*100/int(total_casos_edad()), 2))
-
-
-def fecha_actualizacion_edad():
-    actual_day = date.today()
-    dias_antes = 0
-    day_before = timedelta(days=dias_antes)
-
-    # La primera fecha de la que se tiene datos
-    while df_edad.loc[df_edad['fecha'] == str(actual_day - day_before)].empty:
-        dias_antes += 1
-        day_before = timedelta(days=dias_antes)
-
-    return str(actual_day - day_before)
-
-
-def muertes_totales_espana():
-    return str(df_ccaa_muertes.groupby(['CCAA'])['Fallecidos'].sum()['España'])
-
-
-def muertes_ultimo_dia_espana():
-    return str(muertes_ultimo_dia('España'))
-
-
-def media_muertes_semana_espana():
-    return media_muertes_semana('España')
-
-
-def tasa_letalidad_espana():
-    return str(int(muertes_totales('España'))*100/int(casos_acumulados_espana()))
-
-
-def top_5_casos():
-    ccaa = ['Andalucía', 'Aragón', 'Asturias', 'Baleares', 'C. Valenciana', 'Canarias', 'Cantabria',
-                  'Castilla La Mancha', 'Castilla y León', 'Cataluña', 'Ceuta', 'Extremadura', 'Galicia', 'La Rioja',
-                  'Madrid', 'Melilla', 'Murcia', 'Navarra', 'País Vasco']
-
-    dict = {}
-
-    for i in range(len(ccaa)):
-        dict[ccaa[i]] = int(casos_acumulados(ccaa[i]))
-
-    text=''
-    new_dict = sorted(dict.items(), key=lambda x: x[1], reverse=True)
-
-    for i in range(5):
-        text += '\t - ' + new_dict[i][0] + ': <b>' + str(new_dict[i][1]) + '</b>. \n'
-
-    return text
-
-
-def top_5_casos_100():
-    ccaa = ['Andalucía', 'Aragón', 'Asturias', 'Baleares', 'C. Valenciana', 'Canarias', 'Cantabria',
-                  'Castilla La Mancha', 'Castilla y León', 'Cataluña', 'Ceuta', 'Extremadura', 'Galicia', 'La Rioja',
-                  'Madrid', 'Melilla', 'Murcia', 'Navarra', 'País Vasco']
-
-    dict = {}
-
-    for i in range(len(ccaa)):
-        df_loc = df_ccaa_habitantes.loc[df_ccaa_habitantes['ccaa'] == ccaa[i]]
-        dict[ccaa[i]] = round((int(casos_acumulados(ccaa[i]))*100000)/df_loc['habitantes'].values[0], 1)
-
-    text=''
-    new_dict = sorted(dict.items(), key=lambda x: x[1], reverse=True)
-
-    for i in range(5):
-        text += '\t - ' + new_dict[i][0] + ': <b>' + str(new_dict[i][1]) + '</b>. \n'
-
-    return text
-
-
-def top_5_casos_100_semana():
-    ccaa = ['Andalucía', 'Aragón', 'Asturias', 'Baleares', 'C. Valenciana', 'Canarias', 'Cantabria',
-                  'Castilla La Mancha', 'Castilla y León', 'Cataluña', 'Ceuta', 'Extremadura', 'Galicia', 'La Rioja',
-                  'Madrid', 'Melilla', 'Murcia', 'Navarra', 'País Vasco']
-
-    dict = {}
-
-    for i in range(len(ccaa)):
-        df_loc = df_ccaa_habitantes.loc[df_ccaa_habitantes['ccaa'] == ccaa[i]]
-        dict[ccaa[i]] = round((float(media_casos_semana(ccaa[i]))*100000)/df_loc['habitantes'].values[0], 1)
-
-    text=''
-    new_dict = sorted(dict.items(), key=lambda x: x[1], reverse=True)
-
-    for i in range(5):
-        text += '\t - ' + new_dict[i][0] + ': <b>' + str(new_dict[i][1]) + '</b>. \n'
-
-    return text
-
-
-def top_5_muertes_100():
-    ccaa = ['Andalucía', 'Aragón', 'Asturias', 'Baleares', 'C. Valenciana', 'Canarias', 'Cantabria',
-                  'Castilla La Mancha', 'Castilla y León', 'Cataluña', 'Ceuta', 'Extremadura', 'Galicia', 'La Rioja',
-                  'Madrid', 'Melilla', 'Murcia', 'Navarra', 'País Vasco']
-
-    dict = {}
-
-    for i in range(len(ccaa)):
-        df_loc = df_ccaa_habitantes.loc[df_ccaa_habitantes['ccaa'] == ccaa[i]]
-        dict[ccaa[i]] = round((int(muertes_totales(ccaa[i]))*100000)/df_loc['habitantes'].values[0], 1)
-
-    text=''
-    new_dict = sorted(dict.items(), key=lambda x: x[1], reverse=True)
-
-    for i in range(5):
-        text += '\t - ' + new_dict[i][0] + ': <b>' + str(new_dict[i][1]) + '</b>. \n'
-
-    return text
-
-
-def top_5_muertes_100_semana():
-    ccaa = ['Andalucía', 'Aragón', 'Asturias', 'Baleares', 'C. Valenciana', 'Canarias', 'Cantabria',
-                  'Castilla La Mancha', 'Castilla y León', 'Cataluña', 'Ceuta', 'Extremadura', 'Galicia', 'La Rioja',
-                  'Madrid', 'Melilla', 'Murcia', 'Navarra', 'País Vasco']
-
-    dict = {}
-
-    for i in range(len(ccaa)):
-        df_loc = df_ccaa_habitantes.loc[df_ccaa_habitantes['ccaa'] == ccaa[i]]
-        dict[ccaa[i]] = round((float(media_muertes_semana(ccaa[i]))*100000)/df_loc['habitantes'].values[0], 1)
-
-    text=''
-    new_dict = sorted(dict.items(), key=lambda x: x[1], reverse=True)
-
-    for i in range(5):
-        text += '\t - ' + new_dict[i][0] + ': <b>' + str(new_dict[i][1]) + '</b>. \n'
-
-    return text
-
-
-########################  FUNCIONALIDAD GRÁFICAS  ########################
-inicio_mes = ['2020-01-01', '2020-02-01', '2020-03-01', '2020-04-01', '2020-05-01', '2020-06-01', '2020-07-01',
-              '2020-08-01', '2020-09-01', '2020-10-01', '2020-11-01', '2020-12-01']
-
-def grafica_acumulado(current_autonomy):
-    df_casos_ccaa = df_ccaa_casos.loc[df_ccaa_casos['ccaa'] == current_autonomy]
-    df_casos_ccaa['casos_acumulados'] = df_casos_ccaa['num_casos'].cumsum()
-    df_loc = df_casos_ccaa.loc[:, ['fecha', 'casos_acumulados']]
-    df_loc.set_index("fecha", inplace=True)
-
-    autonomy_lower = normalize(current_autonomy).lower()
-
-    fig, ax = plt.subplots(figsize=(12, 6))
-    x = df_loc.index.get_level_values('fecha')
-
-    plt.bar(x, df_loc['casos_acumulados'], alpha=0.5, width=0.5, label=('Nº Casos'))
-
-    ax.set_xticks(inicio_mes)
-    ax.set_xlim('2020-03-01', x[-1])
-    ax.figure.autofmt_xdate()
-
-    plt.title('Casos acumulados en {}'.format(current_autonomy), fontsize=26)
-    ax.set_ylabel('Nº Casos', fontsize=15)
-    plt.savefig('./img_graficas/acumulado_{}.png'.format(autonomy_lower))
-    plt.close()
-
-
-def grafica_incremento(current_autonomy):
-    df_casos_ccaa = df_ccaa_casos.loc[df_ccaa_casos['ccaa'] == current_autonomy]
-    df_casos_ccaa.set_index("fecha", inplace=True)
-
-    autonomy_lower = normalize(current_autonomy).lower()
-
-    fig, ax = plt.subplots(figsize=(12, 6))
-    x = df_casos_ccaa.index.get_level_values('fecha')
-
-    plt.bar(x, df_casos_ccaa['num_casos'], alpha=0.5, width=0.5, label=('Incremento diario'))
-    plt.plot(x, df_casos_ccaa['num_casos'], color='red')
-
-    ax.set_xticks(inicio_mes)
-    ax.set_xlim('2020-03-01', x[-1])
-    ax.figure.autofmt_xdate()
-
-    plt.title('Incremento de casos en {}'.format(current_autonomy), fontsize=26)
-    ax.set_ylabel('Nº Casos', fontsize=15)
-    plt.savefig('./img_graficas/incremento_{}.png'.format(autonomy_lower))
-    plt.close()
-
-
-def grafica_muertes(current_autonomy):
-    df_muertes_ccaa = df_ccaa_muertes.loc[df_ccaa_muertes['CCAA'] == current_autonomy]
-    df_muertes_ccaa.set_index("Fecha", inplace=True)
-
-    autonomy_lower = normalize(current_autonomy).lower()
-
-    fig, ax = plt.subplots(figsize=(12, 6))
-    x = df_muertes_ccaa.index.get_level_values('Fecha')
-
-    plt.bar(x, df_muertes_ccaa['Fallecidos'], alpha=0.5, width=0.7, label=('Fallecimientos diarios'), color='red')
-    plt.plot(x, df_muertes_ccaa['Fallecidos'], color='red')
-
-    ax.set_xticks(inicio_mes)
-    ax.set_xlim('2020-03-01', x[-1])
-    ax.figure.autofmt_xdate()
-
-    plt.title('Evolución de fallecidos en {}'.format(current_autonomy), fontsize=26)
-    ax.set_ylabel('Fallecidos', fontsize=15)
-    plt.savefig('./img_graficas/fallecidos_{}.png'.format(autonomy_lower))
-    plt.close()
-
-
-def grafica_hospitales(current_autonomy):
-    df_hospitales_ccaa = df_ccaa_hospital.loc[df_ccaa_hospital['CCAA'] == current_autonomy]
-    df_hospitales_ccaa['% Camas Ocupadas UCI COVID'] = df_hospitales_ccaa['% Camas Ocupadas UCI COVID'].fillna('0%')
-    df_hospitales_ccaa['% Camas Ocupadas COVID'] = df_hospitales_ccaa['% Camas Ocupadas COVID'].str.replace(r'%$', '')
-    df_hospitales_ccaa['% Camas Ocupadas UCI COVID'] = df_hospitales_ccaa['% Camas Ocupadas UCI COVID'].str.replace(r'%$', '')
-
-    autonomy_lower = normalize(current_autonomy).lower()
-
-    df_hospitales_ccaa.set_index("Fecha", inplace=True)
-
-    fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(12, 8))
-    x = df_hospitales_ccaa.index.get_level_values('Fecha')
-
-    fig.suptitle('Evolución ocupación camas en {}'.format(current_autonomy), fontsize=26)
-
-    ax1.plot(x, df_hospitales_ccaa['% Camas Ocupadas COVID'], color='blue')
-    ax1.set_ylabel('% camas ocupadas', fontsize=15)
-    ax1.set_xticks(inicio_mes)
-    ax1.set_xlim('2020-08-19', x[-1])
-    ax1.figure.autofmt_xdate()
-
-    ax2.plot(x, df_hospitales_ccaa['% Camas Ocupadas UCI COVID'], color='red')
-    ax2.set_ylabel('% camas UCI ocupadas', fontsize=15)
-    ax2.set_xticks(inicio_mes)
-    ax2.set_xlim('2020-08-19', x[-1])
-    ax2.figure.autofmt_xdate()
-
-    plt.savefig('./img_graficas/hospital_{}.png'.format(autonomy_lower))
-    plt.close()
-
-
-def grafica_incremento_espana():
-    df_casos_fecha = df_ccaa_casos.groupby('fecha')['num_casos'].sum().reset_index()
-    df_casos_fecha.set_index("fecha", inplace=True)
-
-    fig, ax = plt.subplots(figsize=(12, 6))
-    x = df_casos_fecha.index.get_level_values('fecha')
-
-    plt.bar(x, df_casos_fecha['num_casos'], alpha=0.5, width=0.5)
-    plt.plot(x, df_casos_fecha['num_casos'], color='red')
-
-    ax.set_xticks(inicio_mes)
-    ax.set_xlim('2020-03-01', x[-1])
-    ax.figure.autofmt_xdate()
-
-    plt.title('Incremento de casos en España', fontsize=26)
-    ax.set_ylabel('Nº Casos', fontsize=15)
-    plt.savefig('./img_graficas/incremento_espana.png')
-    plt.close()
-
-
-def grafica_acumulado_espana():
-    df_casos_fecha = df_ccaa_casos.groupby('fecha')['num_casos'].sum().reset_index()
-    df_casos_fecha['casos_acumulados'] = df_casos_fecha['num_casos'].cumsum()
-    df_loc = df_casos_fecha.loc[:, ['fecha', 'casos_acumulados']]
-    df_loc.set_index("fecha", inplace=True)
-
-    fig, ax = plt.subplots(figsize=(12, 6))
-    x = df_loc.index.get_level_values('fecha')
-
-    plt.bar(x, df_loc['casos_acumulados'], alpha=0.5, width=0.5)
-    plt.plot(x, df_loc['casos_acumulados'], color='red')
-
-    ax.set_xticks(inicio_mes)
-    ax.set_xlim('2020-03-01', x[-1])
-    ax.figure.autofmt_xdate()
-
-    plt.title('Casos acumulados en España', fontsize=26)
-    ax.set_ylabel('Nº Casos', fontsize=15)
-    plt.savefig('./img_graficas/acumulado_espana.png')
-    plt.close()
-
-
-def grafica_muertes_espana():
-    df_muertes = df_ccaa_muertes.loc[df_ccaa_muertes['CCAA'] == 'España']
-    df_muertes.set_index("Fecha", inplace=True)
-
-    fig, ax = plt.subplots(figsize=(12, 6))
-    x = df_muertes.index.get_level_values('Fecha')
-
-    plt.bar(x, df_muertes['Fallecidos'], alpha=0.5, width=0.7, label=('Fallecimientos diarios'), color='red')
-    plt.plot(x, df_muertes['Fallecidos'], color='red')
-
-    ax.set_xticks(inicio_mes)
-    ax.set_xlim('2020-03-01', x[-1])
-    ax.figure.autofmt_xdate()
-
-    plt.title('Evolución de fallecidos en España', fontsize=26)
-    ax.set_ylabel('Fallecidos', fontsize=15)
-    plt.savefig('./img_graficas/fallecidos_espana.png')
-    plt.close()
-
-
-def grafica_casos_comunidad():
-    ccaa = ['Andalucía', 'Aragón', 'Asturias', 'Baleares', 'C. Valenciana', 'Canarias', 'Cantabria',
-                  'Castilla La Mancha', 'Castilla y León', 'Cataluña', 'Ceuta', 'Extremadura', 'Galicia', 'La Rioja',
-                  'Madrid', 'Melilla', 'Murcia', 'Navarra', 'País Vasco']
-
-    dict = {}
-
-    for i in range(len(ccaa)):
-        dict[ccaa[i]] = int(casos_acumulados(ccaa[i]))
-
-    new_dict = sorted(dict.items(), key=lambda x: x[1], reverse=True)
-    df = pd.DataFrame(new_dict, columns=['ccaa', 'n_casos'])
-    df.set_index("ccaa", inplace=True)
-
-    fig, ax = plt.subplots(figsize=(12, 6))
-    x = df.index.get_level_values('ccaa')
-
-    plt.barh(x, df['n_casos'], alpha=0.5, height=0.8)
-
-    ax.figure.autofmt_xdate()
-
-    plt.title('Casos por comunidad', fontsize=26)
-    plt.gca().invert_yaxis()
-    plt.savefig('./img_graficas/casos_comunidad.png')
-    plt.close()
-
-
-def grafica_casos_100():
-    ccaa = ['Andalucía', 'Aragón', 'Asturias', 'Baleares', 'C. Valenciana', 'Canarias', 'Cantabria',
-                  'Castilla La Mancha', 'Castilla y León', 'Cataluña', 'Ceuta', 'Extremadura', 'Galicia', 'La Rioja',
-                  'Madrid', 'Melilla', 'Murcia', 'Navarra', 'País Vasco']
-
-    dict = {}
-
-    for i in range(len(ccaa)):
-        df_loc = df_ccaa_habitantes.loc[df_ccaa_habitantes['ccaa'] == ccaa[i]]
-        dict[ccaa[i]] = round((int(casos_acumulados(ccaa[i])) * 100000) / df_loc['habitantes'].values[0], 1)
-
-    new_dict = sorted(dict.items(), key=lambda x: x[1], reverse=True)
-    df = pd.DataFrame(new_dict, columns=['ccaa', 'n_casos'])
-    df.set_index("ccaa", inplace=True)
-
-    fig, ax = plt.subplots(figsize=(12, 6))
-    x = df.index.get_level_values('ccaa')
-
-    plt.barh(x, df['n_casos'], alpha=0.5, height=0.8)
-
-    ax.figure.autofmt_xdate()
-
-    plt.title('Casos por cada 100mil habitantes', fontsize=26)
-    plt.gca().invert_yaxis()
-    plt.savefig('./img_graficas/casos_100.png')
-    plt.close()
-
-
-def grafica_casos_100_semana():
-    ccaa = ['Andalucía', 'Aragón', 'Asturias', 'Baleares', 'C. Valenciana', 'Canarias', 'Cantabria',
-                  'Castilla La Mancha', 'Castilla y León', 'Cataluña', 'Ceuta', 'Extremadura', 'Galicia', 'La Rioja',
-                  'Madrid', 'Melilla', 'Murcia', 'Navarra', 'País Vasco']
-
-    dict = {}
-
-    for i in range(len(ccaa)):
-        df_loc = df_ccaa_habitantes.loc[df_ccaa_habitantes['ccaa'] == ccaa[i]]
-        dict[ccaa[i]] = round((float(media_casos_semana(ccaa[i])) * 100000) / df_loc['habitantes'].values[0], 1)
-
-    new_dict = sorted(dict.items(), key=lambda x: x[1], reverse=True)
-    df = pd.DataFrame(new_dict, columns=['ccaa', 'n_muertes'])
-    df.set_index("ccaa", inplace=True)
-
-    fig, ax = plt.subplots(figsize=(12, 6))
-    x = df.index.get_level_values('ccaa')
-
-    plt.barh(x, df['n_muertes'], alpha=0.5, height=0.8)
-
-    ax.figure.autofmt_xdate()
-
-    plt.title('Casos última semana por cada 100mil habitantes', fontsize=26)
-    plt.gca().invert_yaxis()
-
-    for index, value in enumerate(df['n_muertes']):
-        plt.text(value, index, str(value))
-
-    plt.savefig('./img_graficas/casos_100_semana.png')
-    plt.close()
-
-
-def grafica_muertes_100():
-    ccaa = ['Andalucía', 'Aragón', 'Asturias', 'Baleares', 'C. Valenciana', 'Canarias', 'Cantabria',
-                  'Castilla La Mancha', 'Castilla y León', 'Cataluña', 'Ceuta', 'Extremadura', 'Galicia', 'La Rioja',
-                  'Madrid', 'Melilla', 'Murcia', 'Navarra', 'País Vasco']
-
-    dict = {}
-
-    for i in range(len(ccaa)):
-        df_loc = df_ccaa_habitantes.loc[df_ccaa_habitantes['ccaa'] == ccaa[i]]
-        dict[ccaa[i]] = round((int(muertes_totales(ccaa[i])) * 100000) / df_loc['habitantes'].values[0], 1)
-
-    new_dict = sorted(dict.items(), key=lambda x: x[1], reverse=True)
-    df = pd.DataFrame(new_dict, columns=['ccaa', 'n_casos'])
-    df.set_index("ccaa", inplace=True)
-
-    fig, ax = plt.subplots(figsize=(12, 6))
-    x = df.index.get_level_values('ccaa')
-
-    plt.barh(x, df['n_casos'], alpha=0.5, height=0.8)
-
-    ax.figure.autofmt_xdate()
-
-    plt.title('Muertes por cada 100mil habitantes', fontsize=26)
-    plt.gca().invert_yaxis()
-    plt.savefig('./img_graficas/muertes_100.png')
-    plt.close()
-
-
-def grafica_muertes_100_semana():
-    ccaa = ['Andalucía', 'Aragón', 'Asturias', 'Baleares', 'C. Valenciana', 'Canarias', 'Cantabria',
-                  'Castilla La Mancha', 'Castilla y León', 'Cataluña', 'Ceuta', 'Extremadura', 'Galicia', 'La Rioja',
-                  'Madrid', 'Melilla', 'Murcia', 'Navarra', 'País Vasco']
-
-    dict = {}
-
-    for i in range(len(ccaa)):
-        df_loc = df_ccaa_habitantes.loc[df_ccaa_habitantes['ccaa'] == ccaa[i]]
-        dict[ccaa[i]] = round((float(media_muertes_semana(ccaa[i])) * 100000) / df_loc['habitantes'].values[0], 1)
-
-    new_dict = sorted(dict.items(), key=lambda x: x[1], reverse=True)
-    df = pd.DataFrame(new_dict, columns=['ccaa', 'n_casos'])
-    df.set_index("ccaa", inplace=True)
-
-    fig, ax = plt.subplots(figsize=(12, 6))
-    x = df.index.get_level_values('ccaa')
-
-    plt.barh(x, df['n_casos'], alpha=0.5, height=0.8)
-
-    ax.figure.autofmt_xdate()
-
-    plt.title('Muertes última semana por cada 100mil habitantes', fontsize=26)
-    plt.gca().invert_yaxis()
-    plt.savefig('./img_graficas/muertes_100_semana.png')
-    plt.close()
-
-
-def grafica_edades():
-    df_mujeres = df_edad.loc[(df_edad['fecha'] == fecha_actualizacion_edad()) & (df_edad['sexo'] == 'mujeres') & (df_edad['rango_edad'] != 'Total')]
-    df_hombres = df_edad.loc[(df_edad['fecha'] == fecha_actualizacion_edad()) & (df_edad['sexo'] == 'hombres') & (df_edad['rango_edad'] != 'Total')]
-    df_mujeres.set_index("rango_edad", inplace=True)
-    df_hombres.set_index("rango_edad", inplace=True)
-
-    fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(13, 6))
-    x = df_mujeres.index.get_level_values('rango_edad')
-
-    fig.suptitle('Casos por edad', fontsize=26)
-
-    ax1.bar(x, df_mujeres['casos_confirmados'], alpha=0.5, width=0.5, label=('Casos mujeres'), color='blue')
-    ax1.bar(x, df_mujeres['fallecidos'], alpha=0.5, width=0.5, label=('Mujeres fallecidas'), color='green')
-    ax1.set_title('Mujeres')
-    ax1.set_ylabel('Nº Casos', fontsize=15)
-    ax1.legend()
-
-    ax2.bar(x, df_hombres['casos_confirmados'], alpha=0.5, width=0.5, label=('Casos hombres'), color='red')
-    ax2.bar(x, df_hombres['fallecidos'], alpha=0.5, width=0.5, label=('Hombres fallecidos'), color='yellow')
-    ax2.set_title('Hombres')
-    ax2.set_ylabel('Nº Casos', fontsize=15)
-    ax2.legend()
-
-    plt.savefig('./img_graficas/casos_edad_espana.png')
-    plt.close()
 
 if __name__ == '__main__':
     main()
